# Force token cache to support audit annotations