@lavalamp 2020-04-17
uh, is it OK for annotations for 1 request to get announced for all of them? what will be in these annotations? Is there a design I can read?