@enj 2020-04-20
> uh, is it OK for annotations for 1 request to get announced for all of them?

Overall, yes, because the assumption is that for some period of time (cache TTL), all requests with the same API audiences and the same bearer token result in the same audit annotations.  This may not be true if the authenticator sets an annotation based on the current time, but that may be okay since cache TTLs are generally small (seconds).

> what will be in these annotations?

That depends on the authenticator but one of the initial use cases is to make it safer to use bound SA tokens (see below).

> Is there a design I can read?

See #89305 and #89549