@lavalamp 2020-04-20
> Overall, yes, because the assumption is that for some period of time (cache TTL), all requests with the same API audiences and the same bearer token result in the same audit annotations. This may not be true if the authenticator sets an annotation based on the current time, but that may be okay since cache TTLs are generally small (seconds).

Can you state this in a comment? It'll be very helpful for future readers.