@enj 2020-04-28
If 20% of successful authentications set audit annotations, there does not seem to be any real difference in the benchmark:

Old:

```
goos: darwin
goarch: amd64
pkg: k8s.io/apiserver/pkg/authentication/token/cache
BenchmarkCacheContentions/Simple/keys=256-12         	 6511776	      1069 ns/op	      54 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=256-12        	10329835	       420 ns/op	     157 B/op	       5 allocs/op
BenchmarkCacheContentions/Simple/keys=4096-12        	 6208036	      1115 ns/op	      53 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=4096-12       	 9752660	       436 ns/op	     157 B/op	       5 allocs/op
BenchmarkCacheContentions/Simple/keys=65536-12       	 4540059	      1729 ns/op	      57 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=65536-12      	 9601804	       508 ns/op	     158 B/op	       5 allocs/op
BenchmarkKeyFunc/has_audiences-12                    	 1000000	      3202 ns/op	      36 B/op	       2 allocs/op
BenchmarkKeyFunc/nil_audiences-12                    	 1270927	      3106 ns/op	      36 B/op	       2 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=1-12         	 2137609	      1867 ns/op	         0.000080 lookups/op	     391 B/op	      16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=16-12        	 1590171	      1926 ns/op	         0.000101 lookups/op	     424 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=256-12       	 1843376	      1768 ns/op	         0.000103 lookups/op	     406 B/op	      16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=1-12         	 1801042	      1835 ns/op	         0.000424 lookups/op	     393 B/op	      16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=16-12        	 2006400	      2092 ns/op	         0.000690 lookups/op	     424 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=256-12       	 1681060	      2201 ns/op	         0.000462 lookups/op	     406 B/op	      16 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=1-12        	 1762971	      2030 ns/op	         0.00231 lookups/op	     427 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=16-12       	 1882524	      1990 ns/op	         0.00218 lookups/op	     427 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=256-12      	 1881532	      1972 ns/op	         0.00216 lookups/op	     440 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=1-12       	 1063195	      3122 ns/op	         0.0170 lookups/op	     444 B/op	      18 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=16-12      	 1800205	      2096 ns/op	         0.0113 lookups/op	     438 B/op	      18 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=256-12     	 1715931	      2233 ns/op	         0.0120 lookups/op	     452 B/op	      18 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12       	   38581	     79135 ns/op	         0.766 lookups/op	    1356 B/op	      36 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=16-12      	  889266	      3613 ns/op	         0.0897 lookups/op	     539 B/op	      20 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=256-12     	  891417	      3542 ns/op	         0.0860 lookups/op	     562 B/op	      20 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/authentication/token/cache	162.014s
```

New:

```
goos: darwin
goarch: amd64
pkg: k8s.io/apiserver/pkg/authentication/token/cache
BenchmarkCacheContentions/Simple/keys=256-12         	 6631792	      1194 ns/op	      62 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=256-12        	10466500	       431 ns/op	     165 B/op	       5 allocs/op
BenchmarkCacheContentions/Simple/keys=4096-12        	 6305998	      1169 ns/op	      61 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=4096-12       	 9734997	       448 ns/op	     165 B/op	       5 allocs/op
BenchmarkCacheContentions/Simple/keys=65536-12       	 4416181	      1859 ns/op	      65 B/op	       1 allocs/op
BenchmarkCacheContentions/Striped/keys=65536-12      	 8990439	       491 ns/op	     167 B/op	       5 allocs/op
BenchmarkKeyFunc/has_audiences-12                    	 1127702	      3170 ns/op	      36 B/op	       2 allocs/op
BenchmarkKeyFunc/nil_audiences-12                    	 1302345	      2798 ns/op	      36 B/op	       2 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=1-12         	 2048858	      1789 ns/op	         0.000086 lookups/op	     422 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=16-12        	 2013312	      1853 ns/op	         0.000077 lookups/op	     425 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=100_threads=256-12       	 2004166	      1800 ns/op	         0.000074 lookups/op	     433 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=1-12         	 1826516	      1881 ns/op	         0.000458 lookups/op	     425 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=16-12        	 1844925	      1860 ns/op	         0.000492 lookups/op	     426 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=500_threads=256-12       	 1986118	      1923 ns/op	         0.000400 lookups/op	     438 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=1-12        	 1436548	      2274 ns/op	         0.00280 lookups/op	     428 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=16-12       	 1634299	      2194 ns/op	         0.00256 lookups/op	     429 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=2500_threads=256-12      	 1613397	      2180 ns/op	         0.00246 lookups/op	     444 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=1-12       	  876624	      3531 ns/op	         0.0199 lookups/op	     423 B/op	      17 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=16-12      	 1694340	      2076 ns/op	         0.0116 lookups/op	     443 B/op	      18 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=12500_threads=256-12     	 1410993	      2464 ns/op	         0.0132 lookups/op	     462 B/op	      18 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=1-12       	   44834	     79455 ns/op	         0.736 lookups/op	    1703 B/op	      37 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=16-12      	  876471	      3557 ns/op	         0.0902 lookups/op	     577 B/op	      20 allocs/op
BenchmarkCachedTokenAuthenticator/tokens=62500_threads=256-12     	  895708	      3461 ns/op	         0.0858 lookups/op	     597 B/op	      20 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/authentication/token/cache	159.949s
```

Delta:

```
name                                                  old time/op     new time/op     delta
CacheContentions/Simple/keys=256-12                      1.07µs ± 0%     1.19µs ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=256-12                      420ns ± 0%      431ns ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Simple/keys=4096-12                     1.11µs ± 0%     1.17µs ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=4096-12                     436ns ± 0%      448ns ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Simple/keys=65536-12                    1.73µs ± 0%     1.86µs ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=65536-12                    508ns ± 0%      491ns ± 0%   ~     (p=1.000 n=1+1)
KeyFunc/has_audiences-12                                 3.20µs ± 0%     3.17µs ± 0%   ~     (p=1.000 n=1+1)
KeyFunc/nil_audiences-12                                 3.11µs ± 0%     2.80µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=1-12         1.87µs ± 0%     1.79µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=16-12        1.93µs ± 0%     1.85µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=256-12       1.77µs ± 0%     1.80µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=1-12         1.83µs ± 0%     1.88µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=16-12        2.09µs ± 0%     1.86µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=256-12       2.20µs ± 0%     1.92µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=1-12        2.03µs ± 0%     2.27µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=16-12       1.99µs ± 0%     2.19µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=256-12      1.97µs ± 0%     2.18µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=1-12       3.12µs ± 0%     3.53µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=16-12      2.10µs ± 0%     2.08µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=256-12     2.23µs ± 0%     2.46µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=1-12       79.1µs ± 0%     79.5µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=16-12      3.61µs ± 0%     3.56µs ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=256-12     3.54µs ± 0%     3.46µs ± 0%   ~     (p=1.000 n=1+1)

name                                                  old alloc/op    new alloc/op    delta
CacheContentions/Simple/keys=256-12                       54.0B ± 0%      62.0B ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=256-12                       157B ± 0%       165B ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Simple/keys=4096-12                      53.0B ± 0%      61.0B ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=4096-12                      157B ± 0%       165B ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Simple/keys=65536-12                     57.0B ± 0%      65.0B ± 0%   ~     (p=1.000 n=1+1)
CacheContentions/Striped/keys=65536-12                     158B ± 0%       167B ± 0%   ~     (p=1.000 n=1+1)
KeyFunc/has_audiences-12                                  36.0B ± 0%      36.0B ± 0%   ~     (all equal)
KeyFunc/nil_audiences-12                                  36.0B ± 0%      36.0B ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=1-12           391B ± 0%       422B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=16-12          424B ± 0%       425B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=256-12         406B ± 0%       433B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=1-12           393B ± 0%       425B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=16-12          424B ± 0%       426B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=256-12         406B ± 0%       438B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=1-12          427B ± 0%       428B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=16-12         427B ± 0%       429B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=256-12        440B ± 0%       444B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=1-12         444B ± 0%       423B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=16-12        438B ± 0%       443B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=256-12       452B ± 0%       462B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=1-12       1.36kB ± 0%     1.70kB ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=16-12        539B ± 0%       577B ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=256-12       562B ± 0%       597B ± 0%   ~     (p=1.000 n=1+1)

name                                                  old allocs/op   new allocs/op   delta
CacheContentions/Simple/keys=256-12                        1.00 ± 0%       1.00 ± 0%   ~     (all equal)
CacheContentions/Striped/keys=256-12                       5.00 ± 0%       5.00 ± 0%   ~     (all equal)
CacheContentions/Simple/keys=4096-12                       1.00 ± 0%       1.00 ± 0%   ~     (all equal)
CacheContentions/Striped/keys=4096-12                      5.00 ± 0%       5.00 ± 0%   ~     (all equal)
CacheContentions/Simple/keys=65536-12                      1.00 ± 0%       1.00 ± 0%   ~     (all equal)
CacheContentions/Striped/keys=65536-12                     5.00 ± 0%       5.00 ± 0%   ~     (all equal)
KeyFunc/has_audiences-12                                   2.00 ± 0%       2.00 ± 0%   ~     (all equal)
KeyFunc/nil_audiences-12                                   2.00 ± 0%       2.00 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=1-12           16.0 ± 0%       17.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=16-12          17.0 ± 0%       17.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-12         16.0 ± 0%       17.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=1-12           16.0 ± 0%       17.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=16-12          17.0 ± 0%       17.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=256-12         16.0 ± 0%       17.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=1-12          17.0 ± 0%       17.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=16-12         17.0 ± 0%       17.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=256-12        17.0 ± 0%       17.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=1-12         18.0 ± 0%       17.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=16-12        18.0 ± 0%       18.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=256-12       18.0 ± 0%       18.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=1-12         36.0 ± 0%       37.0 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=16-12        20.0 ± 0%       20.0 ± 0%   ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=256-12       20.0 ± 0%       20.0 ± 0%   ~     (all equal)

name                                                  old lookups/op  new lookups/op  delta
CachedTokenAuthenticator/tokens=100_threads=1-12           0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=16-12          0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=100_threads=256-12         0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=1-12           0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=16-12          0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=500_threads=256-12         0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=1-12          0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=16-12         0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=2500_threads=256-12        0.00 ± 0%       0.00 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=1-12         0.02 ± 0%       0.02 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=16-12        0.01 ± 0%       0.01 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=12500_threads=256-12       0.01 ± 0%       0.01 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=1-12         0.77 ± 0%       0.74 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=16-12        0.09 ± 0%       0.09 ± 0%   ~     (p=1.000 n=1+1)
CachedTokenAuthenticator/tokens=62500_threads=256-12       0.09 ± 0%       0.09 ± 0%   ~     (p=1.000 n=1+1)
```